requirements.txt
python-telegram-bot==20.0
beautifulsoup4==4.12.3
requests==2.31.0
